requirements.txt
# python-dotenv==1.0.1
# trl==0.8.6
# ruamel.yaml==0.18.6
# transformers==4.40.1
# accelerate
# deepspeed==0.14.2
# sentencepiece
# jsonargparse==4.29.0
# wandb
# 
# lightning
vllm
datasets
hydra-core